+++ libs/base/src/text/mod.rs
@@ -10,6 +10,7 @@
 // #![cfg_attr(feature = "safe_text", forbid(unsafe_code))]
 
 pub mod fmt;
+pub mod str;
 
 crate::items! { // structural access: _mods, _pub_mods, _all
     #[allow(unused)]
@@ -21,7 +22,7 @@ crate::items! { // structural access: _mods, _pub_mods, _all
     }
     mod _pub_mods { #![allow(unused)]
         pub use super::{
-            fmt::_all::*,
+            fmt::_all::*, str::_all::*,
         };
     }
     pub(super) mod _all { #![allow(unused)]