+++ libs/base/src/text/str/reexports.rs
@@ -1,0 +1,14 @@
+// devela_base::text::str:reexports
+//
+//! String related re-exports.
+//
+
+#![allow(unused_imports)]
+
+use crate::{_reexport, TAG_TEXT};
+
+_reexport! { rust: core::str,
+    tag: TAG_TEXT!(),
+    doc: "Parse a value from a string.",
+    FromStr
+}