+++ libs/base_std/src/text/str/mod.rs
@@ -1,0 +1,19 @@
+// devela_base_std::text::str
+//
+//! String types and related functionality.
+//
+
+mod reexports;
+
+crate::items! { // structural access: _mods, _all, _always
+    #[allow(unused)]
+    pub use _mods::*;
+
+    mod _mods { #![allow(unused)]
+        pub use super::reexports::*;
+    }
+    pub(super) mod _all { #![allow(unused)]
+        #[doc(inline)]
+        pub use super::_mods::*;
+    }
+}